Se implementan los metodos readAll,delete y update.

+++ task/src/main/java/com/desafio/task/repository/TaskRepository.java
@@ -2,5 +2,5 @@ package com.desafio.task.repository;
 
 import com.desafio.task.entity.Task;
 
-public interface TaskRepository extends IGenericRepository<Task, Integer>{
+public interface TaskRepository extends IGenericRepository<Task, Long>{
 }
