+++ task/src/main/java/com/desafio/task/service/TaskService.java
@@ -2,5 +2,5 @@ package com.desafio.task.service;
 
 import com.desafio.task.entity.Task;
 
-public interface TaskService extends ICRUD<Task, Integer>{
+public interface TaskService extends ICRUD<Task, Long>{
 }
